+++ pydidas/core/generic_params/generic_params_pyfai.py
@@ -1,6 +1,6 @@
 # This file is part of pydidas.
 #
-# Copyright 2023, Helmholtz-Zentrum Hereon
+# Copyright 2023 - 2024, Helmholtz-Zentrum Hereon
 # SPDX-License-Identifier: GPL-3.0-only
 #
 # pydidas is free software: you can redistribute it and/or modify
@@ -21,7 +21,7 @@ Parameters for pyFAI integrations.
 """
 
 __author__ = "Malte Storm"
-__copyright__ = "Copyright 2023, Helmholtz-Zentrum Hereon"
+__copyright__ = "Copyright 2023 - 2024, Helmholtz-Zentrum Hereon"
 __license__ = "GPL-3.0-only"
 __maintainer__ = "Malte Storm"
 __status__ = "Production"
@@ -44,7 +44,7 @@ GENERIC_PARAMS_PYFAI = {
         "type": str,
         "default": "2theta / deg",
         "name": "Radial unit",
-        "choices": ["Q / nm^-1", "r / mm", "2theta / deg"],
+        "choices": ["Q / nm^-1", "Q / A^-1", "r / mm", "2theta / deg", "2theta / rad"],
         "unit": "",
         "allow_None": False,
         "tooltip": "The unit and type of the azimuthal profile.",