+++ CHANGELOG.rst
@@ -11,6 +11,7 @@ Improvements
 - Changed a number of filenames and paths (mainly in the documentation) to 
   reduce the total length of the file names.
 - Updated files to new black 2024 style.
+- Added pyFAI units for 'q / A^-1' and '2theta / rad'.
 
 Bugfixes
 --------